Implementation of response and exception handling
This implements the response and exception handling explained in the README.
This provides users of AndSon the ease of getting the response data (in general,
this is what is typically wanted when a request is called) and having
exceptions thrown on non-200 responses (when response data is most likely not
available). It also provides the ability to get the response object (by passing
a block) and allowing the user to analyze the response object and do advanced
behavior with it.

This doesn't include tests because I wanted to modify how we implemented some of
the system tests before testing this.