Added `params` chainable method for defaulting params for all requests
This adds the `params` method as a chainable method for the client. It can be
called to setup the default params used for all requests. This allows a base
client to be setup with params that are desired in all requests. Then the
when using `call` only additional params have to be passed.