Add testing pattern to disable actually calling requests
This adds a testing pattern to AndSon where responses to
requests can be configured to prevent AndSon clients from
actually making requests. Instead, it will look at what's
configured and return that. This is only turned on if the
env variable `ANDSON_TESTING` is set. The point of this is
to allow AndSon to be put into a testing mode so that it
can be used without having to start up a server and run
requests against that.