Stop confusing exception when AndSon can't bind to a server
Previously, we were getting an `NoMethodError` for `close` on the
connection (because it could never make a connection) when a
client couldn't bind. This fixes that by making sure protocol
connection has been set before trying to close it. With this, you
will get an exception like: `Errno::ECONNREFUSED`, which is
the normal exception for the problem.

Fixes #13