Raise custom "Connection Closed" exception
This is to provide a less confusing exception when the remote
server closes a connection and no response is written. This can
happen when a load balancer is put in front of a Sanford server
and the Sanford server is down. Previously, you would get a
"Protocol version mismatch" error which is confusing and not
actually the problem. This will now say accurately that the
"Connection was closed".

Closes #16