Remove calling services with a version
This removes calling services with a version. Services are no
longer added with a version in Sanford and the protocol no longer
supports passing a version. If "versioned" services are wanted,
the version identifier should be added to the service's name.
This is more flexible and makes it more clear that services can
be versioned independently.