Add comparison methods and call runner attributes to `TestClient`
This adds comparison methods to `AndSonClient`, `TestClient` and
`CallRunner`. This allows comparing if 2 clients or call runners
are equal to each other. This also adds call runner attributes to
the `TestClient`. These were missed when originally implementing
the `TestClient`. This allows the `TestClient` to store params,
a logger and a timeout like a regular client/call runner.