Update system tests, move to our latest conventions
This is a minor change that simply updates the system tests to
our latest conventions. This renames the file to match the part
of `AndSon` that we are system testing. This is our latest
convention for defining system tests. This also updates the tests
to our latest standards. Specifically this included renaming the
support `FakeServer` to `TestServer`. This is to avoid confusion
with the word "fake" in a test suite. The `TestServer` is a
`TCPServer` and processes requests. It is not a full featured
`Sanford` server. This is setup for future changes that need to
add or change system tests.