Update test client to encode requests
This updates the test client to encode requests when its called.
This is a test helper to help ensure that the request is valid and
work with the real client. The test client will now build a request
and encode it. This encoded request is thrown away and not used,
its simply to make sure it can be encoded. This doesn't change any
other functionality.

This also fixes a minor bug in the system tests. Sometimes, opening
a TCP socket when stopping the test server would hang. This updates
it to manually build a TCP socket and use `connect_nonblock`. This
keeps it from blocking but furthermore allows using `IO.select`
with a timeout. This ensures if the process hangs that we can time
it out and move on. We open a connection on the server to try and
ensure we shutdown the thread correctly. The server is simple and
uses `accept` which will block the thread, keeping it from shutting
down if a client socket doesn't connect.